Create SignIn page, refactor styles, make styles and components sharing common logic refs: #ZN-8
