--- app/frontend/javascript/components/authorization/sign_in/sign_in.jsx
@@ -1,15 +1,0 @@
-class SignIn extends React.Component {
-  constructor (props) {
-    super(props);
-  }
-
-  render () {
-    return (
-      <div className="container">
-        Sign in page
-      </div>
-    )
-  }
-}
-
-export default SignIn;
