+++ app/frontend/stylesheets/application.css.scss
@@ -31,4 +31,4 @@
 @import "./components/map/map.scss";
 @import "./components/modals/modal.scss";
 @import "./components/navbar.scss";
-@import "./components/auth/sign_up.scss";
+@import "./components/auth.scss";
