--- app/frontend/stylesheets/components/auth/sign_up.scss
@@ -1,21 +1,0 @@
-#sign-up-container {
-  margin-top: 20vh;
-  margin-left: auto;
-  margin-right: auto;
-
-  .form-actions {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-
-    #already-registered-link {
-      margin-bottom: 4px;
-      margin-top: 8px;
-    }
-  }
-
-  #sign-up-title {
-    margin-bottom: 35px;
-    text-align: center;
-  }
-}
