+++ config/locales/en/auth.en.yml
@@ -1,11 +1,19 @@
 en:
   auth:
     buttons:
+      already_registered: Already have an account?
+      create_account: Create an account
       sign_in: Sign in
       sign_up: Create account
-      already_registered: Already have an account?
-    sign_up:
-      fields:
+    fields:
+      sign_in:
+        errors:
+          password_length: Invalid password.
+          invalid_creds: Incorrect username or password.
+        labels:
+          username: Username
+          password: Password
+      sign_up:
         errors:
           password_length: Password should have at least 6 symbols.
           password_confirm: Passwords should match.
@@ -14,4 +22,6 @@ en:
           password: Password
           password_confirmation: Confirm password
     titles:
+      sign_in: Sign in to ZivertNet
       sign_up: Create your account
+      join_zivertnet: Join ZivertNet
