+++ spec/controllers/api/v1/devise_custom/sessions_controller_spec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+describe Api::V1::DeviseCustom::SessionsController, type: :controller do
+end
